tests/provisioning/requirements.txt
google-cloud-container>=1.0.0
google-cloud-error-reporting>=0.34.0
google-api-core>=1.0.0
requests>=2.24.0
parameterized>=0.7.4
